feat(documentValidatorBlock): merge requestValidationAddon, remove standalone block
