+++ frontend/src/app/modules/policy-engine/policy-engine.module.ts
@@ -44,7 +44,7 @@ import { HttpRequestConfigComponent } from './policy-configuration/blocks/main/h
 import { PolicyTreeComponent } from './policy-configuration/policy-tree/policy-tree.component';
 import { ModulePropertiesComponent } from './policy-configuration/module-properties/module-properties.component';
 import { ModuleComponent } from './policy-configuration/blocks/module/module.component';
-import { ToolComponent } from "./policy-configuration/blocks/tool/tool.component";
+import { ToolComponent } from './policy-configuration/blocks/tool/tool.component';
 import { CreateTokenConfigComponent } from './policy-configuration/blocks/tokens/create-token-config/create-token-config.component';
 import { SwitchConfigComponent } from './policy-configuration/blocks/main/switch-config/switch-config.component';
 import { CommonPropertyComponent } from './policy-configuration/common-property/common-property.component';
@@ -53,7 +53,6 @@ import { PolicySettingsComponent } from './policy-configuration/policy-settings/
 import { ButtonBlockAddonComponent } from './policy-viewer/blocks/button-block-addon/button-block-addon.component';
 import { DropdownBlockAddonComponent } from './policy-viewer/blocks/dropdown-block-addon/dropdown-block-addon.component';
 import { RequestAddonConfigComponent } from './policy-configuration/blocks/documents/request-addon-config/request-addon-config.component';
-import { RequestValidationAddonConfigComponent } from './policy-configuration/blocks/documents/request-validation-addon-config/request-validation-addon-config.component';
 import { RequestDocumentBlockAddonComponent } from './policy-viewer/blocks/request-document-block-addon/request-document-block-addon.component';
 //Viewer
 import { DialogBlock } from './policy-viewer/dialog-block/dialog-block.component';
@@ -165,18 +164,18 @@ import { MintRequestsComponent } from './policy-viewer/mint-requests/mint-reques
 import { TransformationButtonBlockComponent } from './policy-viewer/blocks/transformation-button-block/transformation-button-block.component';
 import { IntegrationButtonBlockComponent } from './policy-viewer/blocks/integration-button-block/integration-button-block.component';
 import { RestoreSavepointDialog } from './policy-viewer/dialogs/restore-savepoint-dialog/restore-savepoint-dialog.component';
-import { AddSavepointDialog } from "./policy-viewer/dialogs/add-savepoint-dialog/add-savepoint-dialog.component";
-import { OnLoadSavepointDialog } from "./policy-viewer/dialogs/on-load-savepoint-dialog/on-load-savepoint-dialog.component";
-import { IgnoreRulesDialog } from "./dialogs/ignore-rules-dialog/ignore-rules-dialog.component";
+import { AddSavepointDialog } from './policy-viewer/dialogs/add-savepoint-dialog/add-savepoint-dialog.component';
+import { OnLoadSavepointDialog } from './policy-viewer/dialogs/on-load-savepoint-dialog/on-load-savepoint-dialog.component';
+import { IgnoreRulesDialog } from './dialogs/ignore-rules-dialog/ignore-rules-dialog.component';
 import { PolicyRepositoryComponent } from './policy-repository/policy-repository.component';
 import { WipeConfigComponent } from './policy-configuration/blocks/tokens/wipe-config/wipe-config.component';
 import { PublishToolDialog } from './dialogs/publish-tool-dialog/publish-tool-dialog.component';
 import { SaveToolDialog } from './dialogs/save-tool-dialog/save-tool-dialog.component';
 import { UserPolicyDialog } from './dialogs/user-policy-dialog/user-policy-dialog.component';
-import { GlobalEventsWriterBlockComponent} from "./policy-viewer/blocks/global-events-writer-block/global-events-writer-block.component";
-import { GlobalEventsReaderBlockComponent } from "./policy-viewer/blocks/global-events-reader-block/global-events-reader-block.component";
-import { GlobalEventsReaderFiltersDialogComponent } from "./policy-viewer/dialogs/global-events-reader-filters-dialog/global-events-reader-filters-dialog.component";
-import { AddGlobalEventTopicDialogComponent } from "./policy-viewer/dialogs/add-global-event-topic/add-global-event-topic-dialog.component";
+import { GlobalEventsWriterBlockComponent} from './policy-viewer/blocks/global-events-writer-block/global-events-writer-block.component';
+import { GlobalEventsReaderBlockComponent } from './policy-viewer/blocks/global-events-reader-block/global-events-reader-block.component';
+import { GlobalEventsReaderFiltersDialogComponent } from './policy-viewer/dialogs/global-events-reader-filters-dialog/global-events-reader-filters-dialog.component';
+import { AddGlobalEventTopicDialogComponent } from './policy-viewer/dialogs/add-global-event-topic/add-global-event-topic-dialog.component';
 import { MathConfigComponent } from './policy-configuration/blocks/calculate/math-config/math-config.component';
 import { MathEditorDialogComponent } from './dialogs/math-editor-dialog/math-editor-dialog.component';
 import { FieldLinkDialog } from './dialogs/field-link-dialog/field-link-dialog.component';
@@ -303,7 +302,6 @@ import { PolicyTestAutomationPopupComponent } from './policy-viewer/policy-test-
         ButtonBlockAddonComponent,
         DropdownBlockAddonComponent,
         RequestAddonConfigComponent,
-        RequestValidationAddonConfigComponent,
         RequestDocumentBlockAddonComponent,
         RequestDocumentBlockDialog,
         DataTransformationConfigComponent,
