+++ frontend/src/app/modules/policy-engine/services/blocks-information.ts
@@ -61,7 +61,6 @@ import { MathConfigComponent } from '../policy-configuration/blocks/calculate/ma
 import { GlobalEventsReaderBlockComponent } from '../policy-viewer/blocks/global-events-reader-block/global-events-reader-block.component';
 import { GlobalEventsWriterBlockComponent } from "../policy-viewer/blocks/global-events-writer-block/global-events-writer-block.component";
 import { IpfsTransformationUIAddonCode } from '../policy-viewer/code/ipfs-transformation-ui-addon';
-import { RequestValidationAddonConfigComponent } from '../policy-configuration/blocks/documents/request-validation-addon-config/request-validation-addon-config.component';
 
 const Container: IBlockSetting = {
     type: BlockType.Container,
@@ -308,19 +307,9 @@ const RequestBlockAddon: IBlockSetting = {
     property: RequestAddonConfigComponent,
     code: null,
     allowedChildren: [{
-        type: BlockType.RequestValidationAddon,
+        type: BlockType.DocumentValidatorBlock,
         group: BlockGroup.UnGrouped,
     }],
-}
-
-const RequestValidationAddon: IBlockSetting = {
-    type: BlockType.RequestValidationAddon,
-    icon: BlockIcons[BlockType.RequestValidationAddon],
-    group: BlockGroup.Documents,
-    header: BlockHeaders.Addons,
-    factory: null,
-    property: RequestValidationAddonConfigComponent,
-    code: null,
 }
 
 const Switch: IBlockSetting = {
@@ -975,5 +964,4 @@ export default [
     HttpRequestUIAddon,
     TransformationUIAddon,
     IpfsTransformationUIAddon,
-    RequestValidationAddon
 ];
