+++ policy-service/src/policy-engine/block-validators/block-validator.ts
@@ -55,7 +55,6 @@ import { ExtractDataBlock } from './blocks/extract-data.js';
 import { ButtonBlockAddon } from './blocks/button-block-addon.js';
 import { DropdownBlockAddon } from './blocks/dropdown-block-addon.js';
 import { RequestVcDocumentBlockAddon } from './blocks/request-vc-document-block-addon.js';
-import { RequestValidationAddon } from './blocks/request-validation-addon.js';
 import { DataTransformationAddon } from './blocks/data-transformation-addon.js';
 import { TransformationButtonBlock } from './blocks/transformation-button-block.js';
 import { IntegrationButtonBlock } from './blocks/integration-button-block.js';
@@ -118,7 +117,6 @@ export const validators = [
     ButtonBlockAddon,
     DropdownBlockAddon,
     RequestVcDocumentBlockAddon,
-    RequestValidationAddon,
     DataTransformationAddon,
     HttpRequestUIAddon,
     TransformationUIAddon,
